Add functionality to manage disabled days
